sort database records and fix show command

+++ src/database/DBAdmin.java
@@ -4,6 +4,8 @@ import commands.AirplaneRecord;
 
 import java.io.*;
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
 
 public class DBAdmin {
     private String database;
@@ -22,16 +24,44 @@ public class DBAdmin {
         this.database = database;
     }
 
-    // Serialization:
-    public void saveNewRecord(AirplaneRecord record) {
+    private boolean isExistingDatabase() {
         try {
             FileInputStream fileStream = new FileInputStream(this.database);
-            ObjectInputStream is = new ObjectInputStream(fileStream);
-            Object obj = is.readObject();
-            ArrayList<AirplaneRecord> records = (ArrayList<AirplaneRecord>) obj;
+        } catch(Exception ex) {
+            return false;
+        }
+        return true;
+    }
+
+    // Serialization:
+    public void saveNewRecord(AirplaneRecord record) {
+        if(isExistingDatabase()) {
+            try {
+                FileInputStream fileStream = new FileInputStream(this.database);
+                ObjectInputStream is = new ObjectInputStream(fileStream);
+                Object obj = is.readObject();
+                ArrayList<AirplaneRecord> records = (ArrayList<AirplaneRecord>) obj;
+                try {
+                    FileOutputStream fs = new FileOutputStream(this.database);
+                    ObjectOutputStream os = new ObjectOutputStream(fs);
+                    records.add(record);
+                    Collections.sort(records);
+                    os.writeObject(records);
+                    os.close();
+                    System.out.println("Airplane saved successfully.");
+                } catch (IOException ioException) {
+                    ioException.printStackTrace();
+                }
+                is.close();
+            } catch (Exception ioException) {
+                ioException.printStackTrace();
+            }
+        }
+        else {
             try {
                 FileOutputStream fs = new FileOutputStream(this.database);
                 ObjectOutputStream os = new ObjectOutputStream(fs);
+                ArrayList<AirplaneRecord> records = new ArrayList<AirplaneRecord>();
                 records.add(record);
                 os.writeObject(records);
                 os.close();
@@ -39,9 +69,6 @@ public class DBAdmin {
             } catch (IOException ioException) {
                 ioException.printStackTrace();
             }
-            is.close();
-        } catch (Exception ioException) {
-            ioException.printStackTrace();
         }
     }
